bq_prtitions_to_pg/container/requirements.txt
# BigQuery to PostgreSQL Partition Transfer Dependencies

# Google Cloud Libraries
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0

# PostgreSQL
psycopg2-binary>=2.9.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Configuration
PyYAML>=6.0

# Date/Time handling
python-dateutil>=2.8.0 